requirements.txt
Django==2.0.7
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.7.5
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2018.5